implemented externally-facing capmix module

+++ src/c/lib/memory.h
@@ -17,7 +17,7 @@ typedef struct memory_coord {
 
 capmix_Coord  capmix_addr_coord(capmix_Addr addr);
 void          capmix_memory_init();
-int           capmix_memory_erase(capmix_Addr addr, size_t len);
+int           capmix_memory_erase(capmix_Addr addr, int len);
 uint8_t *     capmix_memory_get(capmix_Addr addr);
-int           capmix_memory_set(capmix_Addr addr, uint8_t *data, size_t len);
+int           capmix_memory_set(capmix_Addr addr, uint8_t *data, int len);
